gogo

+++ app/src/main/java/com/murallaromana/dam/model/entities/Pelicula.kt
@@ -3,7 +3,7 @@ package com.murallaromana.dam.model.entities
 import java.io.Serializable
 
 class Pelicula(
-    var puntuacion: Long,
+    var puntuacion: Double,
     var nombreComun: String,
     var titulo: String,
     var genero: String,
